Ran formatter
